+++ src/app/AppActions.js
@@ -1,0 +1,2 @@
+export const CONFIG_LOADED = 'CONFIG_LOADED';
+export const DATA_CONFIG_LOADED = 'DATA_CONFIG_LOADED';
