+++ src/auth/AuthReducer.js
@@ -1,4 +1,4 @@
-import * as types from '../actions';
+import * as types from 'auth/AuthActions';
 
 const initialState = {
   loading: false,
