+++ webpack.config.dev.js
@@ -13,7 +13,10 @@ module.exports = {
     publicPath: '/'
   },
   resolve: {
-    root: [path.resolve(__dirname, 'src')],
+    root: [
+      path.resolve(__dirname, 'src'),
+      path.resolve(__dirname, 'css')
+    ],
     extensions: ['', '.js', '.jsx']
   },
   plugins: [
